debug : take care of the return type of sqlSearchMethod

+++ springbook.learningtest.spring/src/main/java/springbook/learningtest/spring/jdbc/test-member.sql
@@ -1,5 +1,5 @@
 CREATE TABLE MEMBER (
     id INT PRIMARY KEY,
     name VARCHAR(100),
-    point FLOAT
+    point DOUBLE
 );
